[CS-81] remove WorkMode constraint from Diffusion
Since a lot of legit constraints remain (mockable stuff, logging, etc)
we still define a DiffusionWorkMode type. It's an improvement, because
unlike WorkMode it doesn't offer any logic-layer functionality.

Also extended the logic layer interface with getBlockHeaders,
securityParams, and recoveryInProgress, so satisfy existing
implementations. The latter two were previously delivered by a reader
context, and the first one is a little too complicated for the diffusion
layer to define in terms of getBlock and getBlockHeader.